独立tool call
