+++ config.json
@@ -7,9 +7,9 @@
     "log_level": "DEBUG"
   },
   "api": {
-    "api_key": " ",
-    "base_url": "https://phapi.furina.junmatec.cn/v1",
-    "model": "TIG-3.6-VL-Lite",
+    "api_key": "",
+    "base_url": "https://api.deepseek.com/v1",
+    "model": "deepseek-chat",
     "temperature": 0.7,
     "max_tokens": 2000,
     "max_history_rounds": 10
@@ -25,10 +25,10 @@
     "enabled": true,
     "auto_extract": true,
     "context_length": 5,
-    "similarity_threshold": 0.6,
+    "similarity_threshold": 0.3,
     "neo4j_uri": "neo4j://127.0.0.1:7687",
     "neo4j_user": "neo4j",
-    "neo4j_password": " ",
+    "neo4j_password": "",
     "neo4j_database": "neo4j"
   },
   "handoff": {
@@ -40,19 +40,19 @@
     "playwright_headless": false
   },
   "tts": {
-    "api_key": "your_tts_api_key_here",
+    "api_key": "",
     "port": 5050,
-    "default_voice": "en-US-AvaNeural",
+    "default_voice": "zh-CN-XiaoxiaoNeural",
     "default_format": "mp3",
     "default_speed": 1.0,
-    "default_language": "en-US",
+    "default_language": "zh-CN",
     "remove_filter": false,
     "expand_api": true,
     "require_api_key": true
   },
   "quick_model": {
     "enabled": true,
-    "api_key": " ",
+    "api_key": "",
     "base_url": "https://ark.cn-beijing.volces.com/api/v3",
     "model_name": "doubao-seed-1-6-flash-250615",
     "max_tokens": 512,
@@ -60,21 +60,21 @@
     "timeout": 5,
     "max_retries": 2
   },
-  "weather_api_key": " ",
+  "weather_api_key": "",
   "mqtt": {
     "enabled": true,
-    "broker": "iot-06z00b3k4pfqzty.mqtt.iothub.aliyuncs.com",
+    "broker": "broker.emqx.io",
     "port": 1883,
-    "topic": " ",
-    "client_id": " ",
-    "username": " ",
-    "password": " ",
+    "topic": "device/switch",
+    "client_id": "mcp_device_switch",
+    "username": "",
+    "password": "",
     "keepalive": 60,
     "qos": 1
   },
   "ui": {
     "user_name": "用户",
-    "bg_alpha": 0.5,
+    "bg_alpha": 0.7,
     "window_bg_alpha": 110,
     "mac_btn_size": 36,
     "mac_btn_margin": 16,
